requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coloredlogs==10.0
humanfriendly==4.18
idna==2.8
numpy==1.16.2
opencv-contrib-python==4.1.0.25
Pillow==6.0.0
PyAudio==0.2.11
pycodestyle==2.5.0
pyreadline==2.1
python-memcached==1.59
python-speech-features==0.6
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SpeechRecognition==3.8.1
Unidecode==1.0.23